Typo in CONTRIBUTING.md (#157)
Hi @cicdguy - this is small issue, but this looks like this file was
copied in all our repositories. Can you merge and propagate the change?
The reason for doing that is that our DOCs builds will treat this typo
as a spell mistake. For now we specifically need to add Github to
inst/WORDLIST file in repositories but it should be resolved by fixing
the typo. Thanks!

Signed-off-by: Marcin <133694481+m7pr@users.noreply.github.com>