Add initial testing depth too (#291)
# Pull Request

Set the initial conditions to in case of error to set some testing
depth.
Key line is 559 where the TESTING_DEPTH is added initially to the
variable that will later be used by the GHA.

I also tried to better reflect what is done by replacing `td` with
`TESTING_DEPTH`.
I wonder if the variable `td` is needed at all.